% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_plottr.r
\name{pptx_plot}
\alias{pptx_plot}
\title{Powerpoint Ready Plot}
\usage{
pptx_plot(plt, pptx_path)
}
\arguments{
\item{plt}{A ggplot or gridExtra plot object.}

\item{pptx_path}{Path to powerpoint presentation to export.}
}
\description{
Powerpoint Ready Plots for Separate-and-Editable Figures Made Easy
}
